Unit tests fixed. BaseViewModelTest deleted.
