+++ app/src/test/java/ar/com/wolox/android/bootstrap/login/LoginViewModelTest.kt
@@ -3,7 +3,6 @@ package ar.com.wolox.android.bootstrap.login
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import ar.com.wolox.android.bootstrap.Constants.INTERNAL_SERVER_ERROR_STATUS_CODE
 import ar.com.wolox.android.bootstrap.Constants.USER_IS_LOGGED_KEY
-import ar.com.wolox.android.bootstrap.base.BaseViewModelTest
 import ar.com.wolox.android.bootstrap.base.MainCoroutineRule
 import ar.com.wolox.android.bootstrap.login.LoginTestConstants.EMPTY_PASSWORD
 import ar.com.wolox.android.bootstrap.login.LoginTestConstants.EMPTY_USERNAME
@@ -15,17 +14,19 @@ import ar.com.wolox.android.bootstrap.login.LoginTestConstants.VALID_USERNAME
 import ar.com.wolox.android.bootstrap.repository.UserRepository
 import ar.com.wolox.android.bootstrap.ui.login.InputError
 import ar.com.wolox.android.bootstrap.ui.login.LoginResponse
-import ar.com.wolox.android.bootstrap.ui.login.LoginView
 import ar.com.wolox.android.bootstrap.ui.login.LoginViewModel
 import ar.com.wolox.android.bootstrap.utils.SharedPreferencesManager
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.runBlocking
 import okhttp3.ResponseBody.Companion.toResponseBody
+import org.junit.After
 import org.junit.Assert
+import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.mockito.Mock
 import org.mockito.Mockito
+import org.mockito.MockitoAnnotations
 import org.mockito.kotlin.times
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.verifyNoInteractions
@@ -33,7 +34,20 @@ import org.mockito.kotlin.whenever
 import retrofit2.Response
 
 @ExperimentalCoroutinesApi
-class LoginViewModelTest : BaseViewModelTest<LoginView, LoginViewModel>() {
+class LoginViewModelTest {
+
+    private lateinit var viewModel: LoginViewModel
+
+    @Before
+    fun setUp() {
+        MockitoAnnotations.openMocks(this)
+        viewModel = getViewModelInstance()
+    }
+
+    @After
+    fun tearDown() {
+        Mockito.validateMockitoUsage()
+    }
 
     @get:Rule
     val coroutineTestRule = MainCoroutineRule()
@@ -47,7 +61,7 @@ class LoginViewModelTest : BaseViewModelTest<LoginView, LoginViewModel>() {
     @Mock
     lateinit var sharedPreferencesManager: SharedPreferencesManager
 
-    override fun getViewModelInstance() = LoginViewModel(repository, sharedPreferencesManager)
+    private fun getViewModelInstance() = LoginViewModel(repository, sharedPreferencesManager)
 
     @Test
     fun `given valid credentials, when the user presses the login button, then the user is logged`() = runBlocking {
